Refactoring according to guidelines
